ptxdist: use a deb822 style ptx.sources file instead of ptx.list
With Debian Trixie the keyrings installed in `/etc/apt/trusted.gpg.d/` are
no longer trusted by default.
Instead they have to be mentioned either in the `deb …` line or in a
deb822-style `.sources` file.
Do the latter since it seems to be the more modern approach.

Since this is not available on `bookworm` (which we still want to support
for now) we have to special-case it.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>